Continue small polish refactors
